[GUI] Move widget.h into widget folder
