+++ game/client/main.lua
@@ -1,9 +1,9 @@
 --
 --  Menus
 --
---dofile("game/client/main_menu.lua")
+dofile("game/client/main_menu.lua")
 dofile("game/client/hud.lua")
 
-game.onClientStartup(function()
-    game.gui.push("hud")
-end)
+--game.onClientStartup(function()
+--    game.gui.push("hud")
+--end)
