+++ src/client/gui/component/component.h
@@ -58,6 +58,7 @@ class Component {
     }
 
     virtual void setPosition(const GuiDimension& position) = 0;
+    virtual void setSize(const GuiDimension& position) = 0;
 
   private:
     bool m_isHidden = false;
